requirements/base.txt
dj-database-url==0.5.0
Django==3.2.4
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1